workspace/pipeline-output-servers/deepseek-v3/mcp_git_repository_manager/refined/requirements.txt
mcp[cli]
gitpython
typing_extensions
fastapi
uvicorn